tst_mousefilterstest: avoid sending MouseMove with no movement
Between Qt 5.11 to Qt 5.14, Qt does not de-duplicate MouseMove events
that doesn't actually move the pointer [1]. With the old code, this
causes the positionChanged event to fire once more than originally
expected. We've sort of worked around this in the test, but Qt 5.15 fix
this behavior [2], which breaks the workaround. So, it's better to fix
the problem at the source, which is not to send the duplicated event in
the first place.

[1] https://github.com/qt/qtbase/commit/a37785ec7638e7485112b87dd7e767881fecc114
[2] https://github.com/qt/qtbase/commit/44112d36c86149f41446bec72c38f82981ff6be6