tst_haptics: fix expected duration in test_custom_play()
For some reason, the test asks for the duration of 1200 and then check
for the duration of 400, which doesn't make sense at all. So just
correct it. The intention of the test seems to check if the duration is
modified from the default of 10, so this should still do the job.

The reason it never caused a problem is that, when there's no Qt haptics
plugin, the dummy one that does nothing is loaded. The test anticipates
for this and sort of skips the test. I'm not sure why it has to be
skipped, but suffice to say, this code path is not tested very often.

Also bump the timeout in waitForHapticsCompleted() to cover this longer
duration.